l1: move send_messages_to_l2 util into `AnvilBaseLayer`
Encodes an `L1HandlerTransaction` so that it can be sent to L1.
Move into a method on `AnvilBaseLayer` for most cases, instead of adding
a method to the auto-generated `Starknet` solidity contract.
Using a free function to allow calling this function on multiple
Starknet contracts (as is the case on sepolia with legacy contracts).
